Fix occasional crash when starting new game.
On starting a new game, bonuses are placed in the arena. To this end,
random bonusses are created and assigend to blocks. However, an
erroneous calculation of the index accessing the m_blocks array led to
occasional segmentation faults (index out of range).

This commit fixes the calculation of the index.

(cherry picked from commit 3599732222665d283cad14ac90fc2dbcfe712b1f)